+++ cfme/tests/test_rest.py
@@ -11,7 +11,7 @@ from utils.providers import setup_a_provider as _setup_a_provider
 from utils.version import current_version
 from utils.virtual_machines import deploy_template
 from utils.wait import wait_for
-from utils import mgmt_system, testgen, conf, version
+from utils import testgen, conf, version
 
 
 pytest_generate_tests = testgen.generate(
@@ -27,49 +27,6 @@ pytestmark = [pytest.mark.ignore_stream("5.3")]
 @pytest.fixture(scope="module")
 def a_provider():
     return _setup_a_provider("infra")
-
-
-@pytest.fixture(scope="module")
-def provision_data(
-        rest_api_modscope, provider, small_template):
-    templates = rest_api_modscope.collections.templates.find_by(name=small_template)
-    for template in templates:
-        if template.ems.name == provider.data["name"]:
-            guid = template.guid
-            break
-    else:
-        raise Exception("No such template {} on provider!".format(small_template))
-    result = {
-        "version": "1.1",
-        "template_fields": {
-            "guid": guid
-        },
-        "vm_fields": {
-            "number_of_cpus": 1,
-            "vm_name": "test_rest_prov_{}".format(fauxfactory.gen_alphanumeric()),
-            "vm_memory": "2048",
-            "vlan": provider.data["provisioning"]["vlan"],
-        },
-        "requester": {
-            "user_name": "admin",
-            "owner_first_name": "John",
-            "owner_last_name": "Doe",
-            "owner_email": "jdoe@sample.com",
-            "auto_approve": True
-        },
-        "tags": {
-            "network_location": "Internal",
-            "cc": "001"
-        },
-        "additional_values": {
-            "request_id": "1001"
-        },
-        "ems_custom_attributes": {},
-        "miq_custom_attributes": {}
-    }
-    if isinstance(provider.mgmt, mgmt_system.RHEVMSystem):
-        result["vm_fields"]["provision_type"] = "native_clone"
-    return result
 
 
 @pytest.mark.usefixtures("logged_in")
@@ -100,37 +57,6 @@ def automation_requests_data():
             "auto_approve": True
         }
     } for index in range(1, 5)]
-
-
-# Here also available the ability to create multiple provision request, but used the save
-# href and method, so it doesn't make any sense actually
-@pytest.mark.meta(server_roles="+automate")
-@pytest.mark.usefixtures("setup_provider")
-def test_provision(request, provision_data, provider, rest_api):
-    """Tests provision via REST API.
-    Prerequisities:
-        * Have a provider set up with templates suitable for provisioning.
-    Steps:
-        * POST /api/provision_requests (method ``create``) the JSON with provisioning data. The
-            request is returned.
-        * Query the request by its id until the state turns to ``finished`` or ``provisioned``.
-    Metadata:
-        test_flag: rest, provision
-    """
-
-    vm_name = provision_data["vm_fields"]["vm_name"]
-    request.addfinalizer(
-        lambda: provider.mgmt.delete_vm(vm_name) if provider.mgmt.does_vm_exist(vm_name) else None)
-    request = rest_api.collections.provision_requests.action.create(**provision_data)[0]
-
-    def _finished():
-        request.reload()
-        if request.status.lower() in {"error"}:
-            pytest.fail("Error when provisioning: `{}`".format(request.message))
-        return request.request_state.lower() in {"finished", "provisioned"}
-
-    wait_for(_finished, num_sec=600, delay=5, message="REST provisioning finishes")
-    assert provider.mgmt.does_vm_exist(vm_name), "The VM {} does not exist!".format(vm_name)
 
 
 def test_provider_refresh(request, a_provider, rest_api):
